fix(model): emit limits and cost on the config channel opencode reads
